feat: Create SeparatorGenerator class
